+++ libraries/exoplayer_dash/src/main/java/androidx/media3/exoplayer/dash/DefaultDashChunkSource.java
@@ -163,6 +163,12 @@ public class DefaultDashChunkSource implements DashChunkSource {
   private boolean missingLastSegment;
 
   /**
+   * The time at which the last {@link #getNextChunk(LoadingInfo, long, List, ChunkHolder)} method
+   * was called, as measured by {@link SystemClock#elapsedRealtime}.
+   */
+  private long lastChunkRequestRealtimeMs;
+
+  /**
    * @param chunkExtractorFactory Creates {@link ChunkExtractor} instances to use for extracting
    *     chunks.
    * @param manifestLoaderErrorThrower Throws errors affecting loading of manifests.
@@ -215,6 +221,7 @@ public class DefaultDashChunkSource implements DashChunkSource {
     this.maxSegmentsPerLoad = maxSegmentsPerLoad;
     this.playerTrackEmsgHandler = playerTrackEmsgHandler;
     this.cmcdConfiguration = cmcdConfiguration;
+    this.lastChunkRequestRealtimeMs = C.TIME_UNSET;
 
     long periodDurationUs = manifest.getPeriodDurationUs(periodIndex);
 
@@ -371,7 +378,6 @@ public class DefaultDashChunkSource implements DashChunkSource {
     long availableLiveDurationUs = getAvailableLiveDurationUs(nowUnixTimeUs, playbackPositionUs);
     trackSelection.updateSelectedTrack(
         playbackPositionUs, bufferedDurationUs, availableLiveDurationUs, queue, chunkIterators);
-
     int selectedTrackIndex = trackSelection.getSelectedIndex();
 
     @Nullable
@@ -382,8 +388,13 @@ public class DefaultDashChunkSource implements DashChunkSource {
                 cmcdConfiguration,
                 trackSelection,
                 bufferedDurationUs,
+                /* playbackRate= */ loadingInfo.playbackSpeed,
                 /* streamingFormat= */ CmcdHeadersFactory.STREAMING_FORMAT_DASH,
-                /* isLive= */ manifest.dynamic);
+                /* isLive= */ manifest.dynamic,
+                /* didRebuffer= */ loadingInfo.rebufferedSince(lastChunkRequestRealtimeMs),
+                /* isBufferEmpty= */ queue.isEmpty());
+    lastChunkRequestRealtimeMs = SystemClock.elapsedRealtime();
+
     RepresentationHolder representationHolder = updateSelectedBaseUrl(selectedTrackIndex);
     if (representationHolder.chunkExtractor != null) {
       Representation selectedRepresentation = representationHolder.representation;