+++ libraries/exoplayer_smoothstreaming/src/main/java/androidx/media3/exoplayer/smoothstreaming/DefaultSsChunkSource.java
@@ -18,6 +18,7 @@ package androidx.media3.exoplayer.smoothstreaming;
 import static androidx.media3.exoplayer.trackselection.TrackSelectionUtil.createFallbackOptions;
 
 import android.net.Uri;
+import android.os.SystemClock;
 import androidx.annotation.Nullable;
 import androidx.media3.common.C;
 import androidx.media3.common.Format;
@@ -99,6 +100,12 @@ public class DefaultSsChunkSource implements SsChunkSource {
   @Nullable private IOException fatalError;
 
   /**
+   * The time at which the last {@link #getNextChunk(LoadingInfo, long, List, ChunkHolder)} method
+   * was called, as measured by {@link SystemClock#elapsedRealtime}.
+   */
+  private long lastChunkRequestRealtimeMs;
+
+  /**
    * @param manifestLoaderErrorThrower Throws errors affecting loading of manifests.
    * @param manifest The initial manifest.
    * @param streamElementIndex The index of the stream element in the manifest.
@@ -119,6 +126,7 @@ public class DefaultSsChunkSource implements SsChunkSource {
     this.trackSelection = trackSelection;
     this.dataSource = dataSource;
     this.cmcdConfiguration = cmcdConfiguration;
+    this.lastChunkRequestRealtimeMs = C.TIME_UNSET;
 
     StreamElement streamElement = manifest.streamElements[streamElementIndex];
     chunkExtractors = new ChunkExtractor[trackSelection.length()];
@@ -290,10 +298,14 @@ public class DefaultSsChunkSource implements SsChunkSource {
                     cmcdConfiguration,
                     trackSelection,
                     bufferedDurationUs,
+                    /* playbackRate= */ loadingInfo.playbackSpeed,
                     /* streamingFormat= */ CmcdHeadersFactory.STREAMING_FORMAT_SS,
-                    /* isLive= */ manifest.isLive)
+                    /* isLive= */ manifest.isLive,
+                    /* didRebuffer= */ loadingInfo.rebufferedSince(lastChunkRequestRealtimeMs),
+                    /* isBufferEmpty= */ queue.isEmpty())
                 .setChunkDurationUs(chunkEndTimeUs - chunkStartTimeUs)
                 .setObjectType(CmcdHeadersFactory.getObjectType(trackSelection));
+    lastChunkRequestRealtimeMs = SystemClock.elapsedRealtime();
 
     out.chunk =
         newMediaChunk(